app/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
aiohttp==3.8.4
websockets==11.0.1
python-dotenv==1.0.0
numpy==1.24.2
pydantic==1.10.7
